Move Filter into a new package
